chore(webpackConfig): 🤖  move webPackConfig to web storybook settings [#170441271]
