--- src/Components/setUikitWebpackSetting/index.ts
@@ -1,3 +1,0 @@
-import setUikitWebpackSetting from "./setUikitWebpackSetting";
-
-export default setUikitWebpackSetting;
